requirements.txt
# Dependências essenciais
pydantic~=2.10.4
pydantic_core~=2.27.2
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
requests>=2.32.3  # Para API do Ollama e compatibilidade com browser-use
toml~=0.10.2     # Para leitura de arquivos de configuração
aiohttp~=3.11.13  # Para chamadas async HTTP

# Ferramentas para web e busca
browsergym~=0.13.3
browser-use~=0.1.40
googlesearch-python~=1.3.0
baidusearch~=1.0.3
duckduckgo_search~=7.5.1
html2text~=2024.2.26

# Ferramentas para frontend web
fastapi~=0.115.11
uvicorn~=0.34.0

# Utilitários diversos
aiofiles~=24.1.0
colorama~=0.4.6
playwright~=1.49.1
pillow~=10.4.0
unidiff~=0.7.5

# Dependência para NumPy (usado por várias bibliotecas)
numpy
